% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.Data.R
\name{example.man.fasta}
\alias{example.man.fasta}
\alias{RNAmodR.Data.man}
\alias{example.man.gff3}
\title{RNAmodR example data for tests and man pages}
\usage{
RNAmodR.Data.man()
}
\description{
This data is only used for tests and man page examples
}
\section{Datasets}{

\subsection{example.man.fasta}{
   sequence of artificial genome for S. cerevisiae containing partial 
   sequences of the 18S rRNA
}
\subsection{example.man.gff3}{
   annotation of artificial genome for S. cerevisiae containing partial 
   sequences of the 18S rRNA
}
}

\examples{
RNAmodR.Data.man()
}
