% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.Data.R
\name{RNAmodR.Data.AAS}
\alias{RNAmodR.Data.AAS}
\alias{RNAmodR.Data.example.AAS.fasta}
\alias{RNAmodR.Data.example.AAS.gff3}
\alias{RNAmodR.Data.example.bud23.1}
\alias{RNAmodR.Data.example.bud23.2}
\alias{RNAmodR.Data.example.trm8.1}
\alias{RNAmodR.Data.example.trm8.2}
\alias{RNAmodR.Data.example.wt.1}
\alias{RNAmodR.Data.example.wt.2}
\alias{RNAmodR.Data.example.wt.3}
\title{RNAmodR AlkAnilineSeq example data}
\description{
This dataset contains exmaple data for AlkAnilineSeq.
The indivicual identifiers are \code{RNAmodR.Data.} plus the header from the
\code{Datasets} section.
}
\section{Datasets}{

\subsection{example.AAS.fasta}{
   sequence of artificial genome for S. cerevisiae containing only the 18S 
   rRNA sequence and 10 tRNA sequences
}
\subsection{example.AAS.gff3}{
   annotation of artificial genome for S. cerevisiae containing only the 18S 
   rRNA sequence and 10 tRNA sequences
}
\subsection{example.bud23.1}{
   sequencing reads mapped to artificial genome from bud23del strain - 
   replicate 1
}
\subsection{example.bud23.2}{
   sequencing reads mapped to artificial genome from bud23del strain - 
   replicate 2
}
\subsection{example.trm8.1}{
   sequencing reads mapped to artificial genome from trm8del strain - 
   replicate 1
}
\subsection{example.trm8.2}{
   sequencing reads mapped to artificial genome from trm8del strain - 
   replicate 2
}
\subsection{example.wt.1}{
   sequencing reads mapped to artificial genome from wild type strain - 
   replicate 1
}
\subsection{example.wt.2}{
   sequencing reads mapped to artificial genome from wild type strain - 
   replicate 2
}
\subsection{example.wt.3}{
   sequencing reads mapped to artificial genome from wild type strain - 
   replicate 3
}
}

\examples{
RNAmodR.Data.example.AAS.fasta()
RNAmodR.Data.example.AAS.gff3()
RNAmodR.Data.example.bud23.1()
RNAmodR.Data.example.bud23.2()
RNAmodR.Data.example.trm8.1()
RNAmodR.Data.example.trm8.2()
RNAmodR.Data.example.wt.1()
RNAmodR.Data.example.wt.2()
RNAmodR.Data.example.wt.3()
}
