% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.Data.R
\name{RNAmodR.Data.example}
\alias{RNAmodR.Data.example}
\alias{example.fasta}
\alias{example.gff3}
\alias{example.bam.1}
\alias{example.bam.2}
\alias{example.bam.3}
\title{RNAmodR general example data}
\usage{
RNAmodR.Data.example()
}
\description{
RNAmodR general example data
}
\section{Datasets}{

\subsection{example.fasta}{
   sequences of artificial genome for S. cerevisiae containing only rRNA and 
   tRNA sequences
}
\subsection{example.gff3}{
   annotation of artificial genome for S. cerevisiae containing only rRNA and 
   tRNA sequences
}
\subsection{example.bam.1}{
   sequencing reads mapped to artificial genome - replicate 1
}
\subsection{example.bam.2}{
   sequencing reads mapped to artificial genome - replicate 2
}
\subsection{example.bam.3}{
   sequencing reads mapped to artificial genome - replicate 3
}
}

\examples{
RNAmodR.Data.example()
}
