% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.Data.R
\name{RNAmodR.Data.RMS}
\alias{RNAmodR.Data.RMS}
\alias{example.RMS.fasta}
\alias{example.RMS.gff3}
\alias{example.RMS.1}
\alias{example.RMS.2}
\title{RNAmodR RiboMethSeq example data}
\usage{
RNAmodR.Data.RMS()
}
\description{
RNAmodR RiboMethSeq example data
}
\section{Datasets}{

\subsection{example.RMS.fasta}{
   sequence of artificial genome for H. sapiens containing only the 5.8S rRNA
   sequence
}
\subsection{example.RMS.gff3}{
   annotation of artificial genome for H. sapiens containing only the 5.8S 
   rRNA sequence
}
\subsection{example.RMS.1}{
   sequencing reads mapped to artificial genome - replicate 1
}
\subsection{example.RMS.2}{
   sequencing reads mapped to artificial genome - replicate 2
}
}

\examples{
RNAmodR.Data.RMS()
}
