% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.Data.R
\name{RNAmodR.Data.snoRNAdb}
\alias{RNAmodR.Data.snoRNAdb}
\title{snoRNAdb data}
\description{
The csv files contains a copy of data from the snoRNAdb 
(\url{https://www-snorna.biotoul.fr/}) downloaded on the 2019-02-11.
The coordinates were updated to current rRNA sequences of hg38.
}
\examples{
RNAmodR.Data.snoRNAdb()
}
