% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.Data.R
\name{example.man.fasta}
\alias{example.man.fasta}
\alias{RNAmodR.Data.example.man.fasta}
\alias{RNAmodR.Data.example.man.gff3}
\title{RNAmodR example data for tests and man pages}
\description{
This dataset contains a small data set for tests and man page examples.
The indivicual identifiers are \code{RNAmodR.Data.} plus the header from the
\code{Datasets} section.
}
\section{Datasets}{

\subsection{example.man.fasta}{
   sequence of artificial genome for S. cerevisiae containing partial 
   sequences of the 18S rRNA
}
\subsection{example.man.gff3}{
   annotation of artificial genome for S. cerevisiae containing partial 
   sequences of the 18S rRNA
}
}

