% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.Data.R
\name{RNAmodR.Data.AAS}
\alias{RNAmodR.Data.AAS}
\alias{example.AAS.fasta}
\alias{example.AAS.gff3}
\alias{example.bud23.1}
\alias{example.bud23.2}
\alias{example.trm8.1}
\alias{example.trm8.2}
\alias{example.trm8.3}
\alias{example.wt.1}
\alias{example.wt.2}
\alias{example.wt.3}
\title{RNAmodR AlkAnilineSeq example data}
\usage{
RNAmodR.Data.AAS()
}
\description{
RNAmodR AlkAnilineSeq example data
}
\section{Datasets}{

\subsection{example.AAS.fasta}{
   sequence of artificial genome for S. cerevisiae containing only the 18S 
   rRNA sequence and 10 tRNA sequences
}
\subsection{example.AAS.gff3}{
   annotation of artificial genome for S. cerevisiae containing only the 18S 
   rRNA sequence and 10 tRNA sequences
}
\subsection{example.bud23.1}{
   sequencing reads mapped to artificial genome from bud23del strain - 
   replicate 1
}
\subsection{example.bud23.2}{
   sequencing reads mapped to artificial genome from bud23del strain - 
   replicate 2
}
\subsection{example.trm8.1}{
   sequencing reads mapped to artificial genome from trm8del strain - 
   replicate 1
}
\subsection{example.trm8.2}{
   sequencing reads mapped to artificial genome from trm8del strain - 
   replicate 2
}
\subsection{example.trm8.3}{
   sequencing reads mapped to artificial genome from trm8del strain - 
   replicate 3
}
\subsection{example.wt.1}{
   sequencing reads mapped to artificial genome from wild type strain - 
   replicate 1
}
\subsection{example.wt.2}{
   sequencing reads mapped to artificial genome from wild type strain - 
   replicate 2
}
\subsection{example.wt.3}{
   sequencing reads mapped to artificial genome from wild type strain - 
   replicate 3
}
}

\examples{
RNAmodR.Data.AAS()
}
