% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.Data.R
\name{RNAmodR.Data.example}
\alias{RNAmodR.Data.example}
\alias{RNAmodR.Data.example.fasta}
\alias{RNAmodR.Data.example.gff3}
\alias{RNAmodR.Data.example.bam.1}
\alias{RNAmodR.Data.example.bam.2}
\alias{RNAmodR.Data.example.bam.3}
\title{RNAmodR general example data}
\description{
This dataset contains general example data used for different purposes.
The indivicual identifiers are \code{RNAmodR.Data.} plus the header from the
\code{Datasets} section.
}
\section{Datasets}{

\subsection{example.fasta}{
   sequences of artificial genome for S. cerevisiae containing only rRNA and 
   tRNA sequences
}
\subsection{example.gff3}{
   annotation of artificial genome for S. cerevisiae containing only rRNA and
   tRNA sequences
}
\subsection{example.bam.1}{
   sequencing reads mapped to artificial genome - replicate 1
}
\subsection{example.bam.2}{
   sequencing reads mapped to artificial genome - replicate 2
}
\subsection{example.bam.3}{
   sequencing reads mapped to artificial genome - replicate 3
}
}

\examples{
RNAmodR.Data.example.fasta()
RNAmodR.Data.example.gff3()
RNAmodR.Data.example.bam.1()
RNAmodR.Data.example.bam.2()
RNAmodR.Data.example.bam.3()
}
