% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.Data.R
\docType{package}
\name{RNAmodR.Data}
\alias{RNAmodR.Data}
\title{RNAmodR.Data}
\description{
RNAmodR.Data contains example data, which is used for vignettes and example
workflows in the \code{RNAmodR} and dependent 
packages.
}
\author{
Felix G.M. Ernst [aut], Denis L.J. Lafontaine [ctb]
}
