% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.Data.R
\name{RNAmodR.Data.RMS}
\alias{RNAmodR.Data.RMS}
\alias{RNAmodR.Data.example.RMS.fasta}
\alias{RNAmodR.Data.example.RMS.gff3}
\alias{RNAmodR.Data.example.RMS.1}
\alias{RNAmodR.Data.example.RMS.2}
\title{RNAmodR RiboMethSeq example data}
\description{
This dataset contains example data for RiboMethSeq.
The indivicual identifiers are \code{RNAmodR.Data.} plus the header from the
\code{Datasets} section.
}
\section{Datasets}{

\subsection{example.RMS.fasta}{
   sequence of artificial genome for H. sapiens containing only the 5.8S rRNA
   sequence
}
\subsection{example.RMS.gff3}{
   annotation of artificial genome for H. sapiens containing only the 5.8S 
   rRNA sequence
}
\subsection{example.RMS.1}{
   sequencing reads mapped to artificial genome - replicate 1
}
\subsection{example.RMS.2}{
   sequencing reads mapped to artificial genome - replicate 2
}
}

\examples{
RNAmodR.Data.example.RMS.fasta()
RNAmodR.Data.example.RMS.gff3()
RNAmodR.Data.example.RMS.1()
RNAmodR.Data.example.RMS.2()
}
